Outline subcommand invocation
Read the configuration file. Parse the modinfo file. Outline how to
invoke subcommands.

Environment variables are looked up whenever and wherever they are
needed. This should be changed so that all environment variables that
are expected to be set are looked up before subcommands are invoked, and
the values passed down as arguments. In other words, access to external
dependencies should happen from main.rs only. This will be fixed in a
future commit.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>